[Sebas] This is just a quick doodle on how to use Lie Algebras
Basically, we create a local map and make updates based on the lie
algebra.
This works because the lie algebra throught the exponential map
generator will always be in the SO(3) group.

We optimize both using this formulation and without and we can see
that properties of rotation matrices are not respected if we do willy
nilly optimization.